refactor(LottoMoney): return 값 계산 double로 하게 변경

+++ src/main/java/com/javabom/lotto/domain/vo/LottoMoney.java
@@ -13,7 +13,7 @@ public class LottoMoney {
     }
 
     public double calculateRateOfReturn(int profit) {
-        int purchaseAmount = (numberOfAutoTicket + numberOfManualTicket) * LOTTO_PRICE;
+        double purchaseAmount = (numberOfAutoTicket + numberOfManualTicket) * LOTTO_PRICE;
 
         return profit / purchaseAmount;
     }
